==> VIDEO AEC/2_Partición_datos_externa.ipynb <==
import pandas as pd
import numpy as np

#Leer la información del documento Excel original
dataFrame = pd.read_csv('../Material/Database.csv', sep=';')
print(dataFrame)

# Partición externa de los datos
from sklearn.model_selection import train_test_split, KFold

#Validación hold-out
# train, test = train_test_split(dataFrame, test_size=0.2, shuffle=True, random_state=42) # hold-out

#Validación cruzada externa de forma manual
kf = KFold(n_splits=5, shuffle=True, random_state=42)
bolsas = kf.split(dataFrame)

for k, (train_fold, test_fold) in enumerate(bolsas):
    train = dataFrame.iloc[train_fold]
    test = dataFrame.iloc[test_fold]

#Obtener cuantas muestras por clase tengo en los datos de train y en los datos de test, luego del split
lab_train, ocur_train = np.unique(train.Class, return_counts=True)
lab_test, ocur_test = np.unique(test.Class, return_counts=True)

print(' --- TRAIN --- \nGlaucoma: ', ocur_train[0], '\nHealthy: ', ocur_train[1])
print(' --- TEST --- \nGlaucoma: ', ocur_test[0], '\nHealthy: ', ocur_test[1])


print(train)

# Aleatorizar los dataframes
#Desordenar los datos aleatoriamente, para hacer más robusto el apartado de datos de train y test
train = train.sample(frac=1)
test = test.sample(frac=1)

print(train)

# Guardado de los modelos
import os
if not os.path.exists('../partitions'):
    os.mkdir('../partitions')

train.to_csv('../partitions/train.csv', sep=';')
test.to_csv('../partitions/test.csv', sep=';')
