==> VIDEO AEC/5_Modelado.ipynb <==
# Cargar los datos de entrenamiento
import numpy as np
train = np.load('../final_features/train.npy')

X_train = train[:,:-1]
y_train = train[:, -1]

print(np.shape(X_train))
print(np.shape(y_train))

# Definir los modelos de clasificación
from sklearn.linear_model import LogisticRegression #Dice Regression pero es un Algoritmo para ser usado por clasificación
from sklearn.neural_network import MLPClassifier

#Estudiar estos dos algoritmos
#random_state es la semilla
#El número de capas o neuronas debe ser menor al número de características
algoritmos = {'LOGR': LogisticRegression(penalty='l2', solver='saga', max_iter=1000, random_state=42),
             'MLP': MLPClassifier(hidden_layer_sizes=[8,4], activation='relu', solver='sgd', batch_size='auto',
                                 learning_rate='adaptive', learning_rate_init=0.01, max_iter=1000, random_state=42)}
#hIPERPARAMETROS INICIALES
#algoritmos = {'LOGR': LogisticRegression(penalty='l2', solver='lbfgs', max_iter=100, random_state=42),
#             'MLP': MLPClassifier(hidden_layer_sizes=[100,], activation='relu', solver='adam', batch_size='auto',
#                                 learning_rate='constant', learning_rate_init=0.001, max_iter=200, random_state=42)}


# Cross-validation interno en k=5 bolsas
from sklearn.model_selection import cross_val_score, KFold

results={}
for nombre, alg in algoritmos.items():
    results[nombre] = cross_val_score(alg, X_train, y_train, cv=KFold(n_splits=5, shuffle=True, random_state=42))
    print(nombre + ':  Accuracy:  %0.4f +/- %0.4f' % (results[nombre].mean(), results[nombre].std()))
    #El porcentaje de número de datos que estoy acertando en las pediciones

# Definimos el modelo definitivo.
algoritmos = {'LOGR': LogisticRegression(penalty='l2', solver='saga', max_iter=1000, random_state=42),
             'MLP': MLPClassifier(hidden_layer_sizes=[8,4], activation='relu', solver='sgd', batch_size='auto',
                                 learning_rate='adaptive', learning_rate_init=0.01, max_iter=1000, random_state=42)}
#Entrenamos todos los datos luego de mejorar nuestros algoritmos
LOGR_definitivo = LOGR.fit(X_train, y_train)
MLP_definitivo = MLP.fit(X_train, y_train)

# Atributos se obtienen durante el entrenamiento
#Sirven para utilizar cuando se tienen más datos y es necesario ajustar los algoritmos de entrenamiento
#Estos como son de MLP servirian para el alg MLP
print('Mínimo error cometido: ', MLP_definitivo.best_loss_)
#Conocer cuantas iteraciones necesita el entrenamiento
print('Número de iteraciones llevadas a cabo: ', MLP_definitivo.n_iter_)

# Guardar modelos
import os
if not os.path.exists('../models'):
    os.mkdir('../models')
    
import pickle
with open('../models/LOGR.pickle', 'wb') as fw:
    pickle.dump(LOGR_definitivo, fw)
with open('../models/MLP.pickle', 'wb') as fw:
    pickle.dump(MLP_definitivo, fw)
